requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
colorama==0.4.6
Django==6.0.3
djangorestframework==3.17.1
fastapi==0.135.1
idna==3.11
iniconfig==2.3.0
# Editable Git install with no remote (lss_clean==0.1.0)
-e d:\python_projects\lss_clean
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
sqlparse==0.5.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
